+++ RebornCore/src/main/java/reborncore/common/powerSystem/PowerSystem.java
@@ -42,7 +42,7 @@ public class PowerSystem {
 	}
 
 	public static String getLocalizedPowerNoSuffix(double power) {
-		return getRoundedString((int) power, "", true);
+		return getRoundedString(power, "", true);
 	}
 
 	public static String getLocalizedPowerNoFormat(double power){
